ON-45486 # Added styles to FormsAppEnvironment
